requirement.txt
tkinter
jcamp
urllib
requests
pillow
matplotlib
joblib